refactor: update imports and improve HTML attribute bindings for better readability and accessibility
