+++ angular/projects/researchdatabox/app-config/src/app/app-config.component.ts
@@ -17,7 +17,8 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-import { Component, ElementRef, Inject, DOCUMENT } from '@angular/core';
+import { Component, ElementRef, Inject } from '@angular/core';
+import { DOCUMENT } from '@angular/common';
 import { FormGroup, Validators, FormBuilder } from '@angular/forms';
 
 import { BaseComponent, UtilityService, LoggerService, TranslationService, AppConfigService } from '@researchdatabox/portal-ng-common';
