+++ angular/projects/researchdatabox/export/src/app/export.component.html
@@ -6,7 +6,7 @@
     <div class="col-xs-10">
       <div class="form-horizontal">
         <div class="form-group">
-          <label class="col-xs-2 control-label">{{ 'export-modified-after' | i18next }}</label>
+          <label class="col-xs-2 control-label" for="after">{{ 'export-modified-after' | i18next }}</label>
           <span class="col-xs-2">
             <input type="text" id="after"
               [bsConfig]="datePickerOpts"
@@ -16,7 +16,7 @@
           </span>
         </div>
         <div class="form-group">
-          <label class="col-xs-2 control-label">{{ 'export-modified-before' | i18next }}</label>
+          <label class="col-xs-2 control-label" for="before">{{ 'export-modified-before' | i18next }}</label>
           <span class="col-xs-2">
             <input type="text" id="before"
               [bsConfig]="datePickerOpts"
@@ -31,7 +31,7 @@
             @for (recExportFormat of getExportFormatNames(); track recExportFormat) {
               <div  >
                 <input [checked]="recExportFormat.checked" (change)="setExportFormat(recExportFormat.id, $event)" id="format-{{recExportFormat.id}}" name="export-format-radio" type="radio" >
-                <label for="format-{{recExportFormat.name}}" >&nbsp;{{ recExportFormat.name }}</label>
+                <label for="format-{{recExportFormat.id}}" >&nbsp;{{ recExportFormat.name }}</label>
               </div>
             }
           </span>
