+++ angular/projects/researchdatabox/local-auth/src/app/local-auth.component.html
@@ -8,13 +8,13 @@
       <div class="panel-body">
         <div class="form-login">
           <form [formGroup]="form" (ngSubmit)='onLogin($event)'>
-            <input type="text" id="username" class="form-control input-sm chat-input" [placeholder]="'local-creds-input-username-label' | i18next" formControlName="username" [attr.aria-label]="'local-creds-input-username-label' | i18next " autocomplete="off"/>
+            <input type="text" id="username" class="form-control input-sm chat-input" [placeholder]="'local-creds-input-username-label' | i18next" formControlName="username" [attr.aria-label]="'local-creds-input-username-label' | i18next" autocomplete="off"/>
             <br/>
-            <input type="password" id="password" class="form-control input-sm chat-input" [placeholder]="'local-creds-input-password-label' | i18next" formControlName="password" [attr.aria-label]="'local-creds-input-password-label' | i18next " autocomplete="off"/>
+            <input type="password" id="password" class="form-control input-sm chat-input" [placeholder]="'local-creds-input-password-label' | i18next" formControlName="password" [attr.aria-label]="'local-creds-input-password-label' | i18next" autocomplete="off"/>
             <br/>
             <div class="wrapper">
               <span class="group-btn">
-                <button type='submit' class="btn btn-primary btn-md {{isLoginDisabled ? 'disabled' : ''}}" (click)="onLogin($event)">{{ 'local-auth-login' | i18next }} <i class="fa fa-sign-in"></i></button>
+                <button type="submit" class="btn btn-primary btn-md" [disabled]="isLoginDisabled" [class.disabled]="isLoginDisabled">{{ 'local-auth-login' | i18next }} <i class="fa fa-sign-in"></i></button>
               </span>
             </div>
             <br/>
