+++ angular/projects/researchdatabox/manage-users/src/app/manage-users.component.html
@@ -195,7 +195,7 @@
           <div class="bg-{{updateDetailsMsgType}} center-block">{{updateDetailsMsg}}</div>
         </div>
         <div class="modal-footer">
-          <button type="button" class="btn btn-primary" form="detailsform" type="submit">{{ 'manage-users-save' | i18next }}</button>
+          <button type="submit" class="btn btn-primary" form="detailsform">{{ 'manage-users-save' | i18next }}</button>
           <button type="button" class="btn btn-default" (click)="hideDetailsModal()">{{ 'manage-users-cancel' | i18next }}</button>
         </div>
       </div>
@@ -318,7 +318,7 @@
           <div class="bg-{{newUserMsgType}} center-block">{{ newUserMsg }}</div>
         </div>
         <div class="modal-footer">
-          <button type="button" class="btn btn-primary" form="newform" type="submit">{{ 'manage-users-save' | i18next }}</button>
+          <button type="submit" class="btn btn-primary" form="newform">{{ 'manage-users-save' | i18next }}</button>
           <button type="button" class="btn btn-default" (click)="hideNewUserModal()">{{ 'manage-users-cancel' | i18next }}</button>
         </div>
       </div>
